Working on documentation of reduce and scan

+++ pctl/example/max.cpp
@@ -22,12 +22,12 @@ using namespace max_ex;
 void ex() {
   
   {
-    parray<long> xs = { 1, 3, 9, 0, 33, 1, 1 };
+    parray<int> xs = { 1, 3, 9, 0, 33, 1, 1 };
     std::cout << "xs\t\t= " << xs << std::endl;
     std::cout << "max(xs)\t= " << max_ex::max(xs) << std::endl;
   }
   
-  parray<parray<long>> xss = { {23, 1, 3}, { 19, 3, 3 }, { 100 } };
+  parray<parray<int>> xss = { {23, 1, 3}, { 19, 3, 3 }, { 100 } };
   
   {
     std::cout << "xss\t\t= " << xss << std::endl;
